+++ src/core/components/sub_components/SubPseudoRendering.cpp
@@ -4,7 +4,7 @@
 
 #include "core/components/sub_components/SubPseudoRendering.hpp"
 
-SubPseudoRendering::SubPseudoRendering(const char input_tile, const unsigned short int input_color)
+SubPseudoRendering::SubPseudoRendering(const char input_tile, const unsigned int input_color)
     : AbsPseudoRendering(input_tile, input_color) {}
 SubPseudoRendering::~SubPseudoRendering() = default;
 
@@ -12,6 +12,6 @@ char SubPseudoRendering::get_tile() const {
   return tile;
 }
 
-unsigned short int SubPseudoRendering::get_color() {
+unsigned int SubPseudoRendering::get_color() {
   return color;
 }
